# docs/swagger.yaml
definitions:
...
  dto.AvailabilitySlot:
    properties:
      day:
        type: string
      endTime:
        type: string
      location:
        type: string
      startTime:
        type: string
    type: object
...
  dto.CreateMatchRequest:
    properties:
      availabilitySlots:
        items:
          $ref: '#/definitions/dto.AvailabilitySlot'
        type: array
      courtId:
        type: string
      matchType:
        type: string
      ntrpMax:
        type: number
      ntrpMin:
        type: number
      participantIds:
        items:
          type: string
        type: array
      playTypes:
        description: rally, singles, doubles,omitempty"`
        items:
          type: string
        type: array
      preferredCourt:
        type: string
      scheduledAt:
        type: string
      specialRequirements:
        type: string
    required:
    - matchType
    - participantIds
    type: object
...
  dto.FindCompetitiveMatchesRequest:
    properties:
      age_range:
        $ref: '#/definitions/services.AgeRange'
      availability:
        items:
          $ref: '#/definitions/services.AvailabilityCriteria'
        type: array
      gender:
        type: string
      limit:
        type: integer
      location:
        $ref: '#/definitions/services.LocationCriteria'
      match_type:
        description: singles, doubles
        type: string
      max_distance:
        type: number
      min_reputation_score:
        description: 較高的信譽要求
        type: number
      ntrp_range:
        allOf:
        - $ref: '#/definitions/services.NTRPRange'
        description: 嚴格的等級匹配
      preferred_court_type:
        type: string
    type: object
...
  dto.FindPartnersRequest:
    properties:
      age_range:
        $ref: '#/definitions/services.AgeRange'
      availability:
        items:
          $ref: '#/definitions/services.AvailabilityCriteria'
        type: array
      gender:
        type: string
      limit:
        type: integer
      location:
        $ref: '#/definitions/services.LocationCriteria'
      max_distance:
        type: number
      min_reputation_score:
        type: number
      ntrp_range:
        allOf:
        - $ref: '#/definitions/services.NTRPRange'
        description: 較寬鬆的範圍
      play_type:
        description: rally, doubles, singles
        items:
          type: string
        type: array
      playing_frequency:
        type: string
    type: object
...
  services.AgeRange:
    properties:
      max:
        type: integer
      min:
        type: integer
    type: object
  services.AvailabilityCriteria:
    properties:
      time:
        description: morning, afternoon, evening
        type: string
      type:
        description: weekday, weekend
        type: string
    type: object
  services.LocationCriteria:
    properties:
      city:
        type: string
      district:
        type: string
    type: object
  services.NTRPRange:
    properties:
      max:
        type: number
      min:
        type: number
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 網球平台後端 API 服務
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 網球平台 API
  version: "1.0"
paths:
...
  /api/v1/matches/create:
    post:
      consumes:
      - application/json
      description: 創建新的競賽對戰配對
      parameters:
      - description: 配對資訊
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateMatchRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 創建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 請求錯誤
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授權
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 伺服器錯誤
          schema:
            additionalProperties: true
            type: object
      summary: 創建競賽配對
      tags:
      - matches
  /api/v1/matches/find:
    post:
      consumes:
      - application/json
      description: 尋找適合對戰的對手，重視技術匹配和信譽分數
      parameters:
      - description: 對手篩選條件
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.FindCompetitiveMatchesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 對手列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 請求錯誤
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授權
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 伺服器錯誤
          schema:
            additionalProperties: true
            type: object
      summary: 尋找對手
      tags:
      - matches
  /api/v1/matches/history:
    get:
      description: 獲取過往的競賽對戰記錄
      parameters:
      - default: 1
        description: 頁碼
        in: query
        name: page
        type: integer
      - default: 10
        description: 每頁數量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 對戰歷史
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授權
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 伺服器錯誤
          schema:
            additionalProperties: true
            type: object
      summary: 獲取對戰歷史
      tags:
      - matches
  /api/v1/partners/create:
    post:
      consumes:
      - application/json
      description: 創建新的球友練習配對
      parameters:
      - description: 配對資訊
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateMatchRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 創建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 請求錯誤
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授權
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 伺服器錯誤
          schema:
            additionalProperties: true
            type: object
      summary: 創建球友練習配對
      tags:
      - partners
  /api/v1/partners/find:
    post:
      consumes:
      - application/json
      description: 尋找適合練習的球友，重視位置接近度和時間彈性
      parameters:
      - description: 球友篩選條件
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.FindPartnersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 球友列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 請求錯誤
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授權
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 伺服器錯誤
          schema:
            additionalProperties: true
            type: object
      summary: 尋找球友
      tags:
      - partners
  /api/v1/partners/history:
    get:
      description: 獲取過往的練習球友記錄
      parameters:
      - default: 1
        description: 頁碼
        in: query
        name: page
        type: integer
      - default: 10
        description: 每頁數量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 球友歷史
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授權
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 伺服器錯誤
          schema:
            additionalProperties: true
            type: object
      summary: 獲取球友歷史
      tags:
      - partners
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"